# docs/specs/swagger.yaml
{ "openapi": "3.0.1","info": { "title": "Fresco","description": "project.description","contact": { "name": "bootcamp12","email": "bootcamp12@mercadolibre.com" },"version": "project.version" },"servers": [ { "url": "https://lalala-integration_test_Fresco.furyapps.io","description": "Scope lalala-integration_test" } ],"security": [ { "bearerAuth": [ ] } ],"paths": { "/api/v1/fresh-products/orders/": { "put": { "tags": [ "Purchase orders" ],"summary": "Update purchase order by id","operationId": "updateOrder","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderDTO" } } },"required": true },"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderDTO" } } } } } },"post": { "tags": [ "Purchase orders" ],"summary": "Create new purchase order","operationId": "create","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderDTO" } } },"required": true },"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"201": { "description": "Created","content": { "application/json": { "schema": { "$ref": "#/components/schemas/PurchaseOrderDTO" } } } } } } },"/api/v1/fresh-products/inboundorder": { "put": { "tags": [ "Warehouse orders" ],"summary": "Update products from inventory","operationId": "update","parameters": [ { "name": "Authorization","in": "header","required": true,"schema": { "type": "string" } } ],"requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboundOrderDTO" } } },"required": true },"responses": { "400": { "description": "Bad Request","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"201": { "description": "Created","content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboundOrderResponseDTO" } } } } } },"post": { "tags": [ "Warehouse orders" ],"summary": "Add products to inventory","operationId": "create_1","parameters": [ { "name": "Authorization","in": "header","required": true,"schema": { "type": "string" } } ],"requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboundOrderDTO" } } },"required": true },"responses": { "400": { "description": "Bad Request","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"201": { "description": "Created","content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboundOrderResponseDTO" } } } } } } },"/api/v1/sign-in": { "post": { "tags": [ "Sessions" ],"summary": "Sign-in to user account","operationId": "login","parameters": [ { "name": "username","in": "query","required": true,"schema": { "type": "string" } },{ "name": "password","in": "query","required": true,"schema": { "type": "string" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponseDTO" } } } } } } },"/api/v1/countries/": { "post": { "tags": [ "Country houses" ],"summary": "Create country house","operationId": "create_2","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CountryHouseDTO" } } },"required": true },"responses": { "400": { "description": "Bad Request","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"201": { "description": "Created","content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboundOrderResponseDTO" } } } } } } },"/ping": { "get": { "tags": [ "Health check" ],"summary": "Check API health","operationId": "ping","responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponseDTO" } } } } } } },"/api/v1/warehouse": { "get": { "tags": [ "Warehouse catalog" ],"summary": "List products by category","operationId": "groupCount","parameters": [ { "name": "productCode","in": "query","required": true,"schema": { "type": "string" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } } } } },"/api/v1/fresh-products": { "get": { "tags": [ "Product catalog" ],"summary": "List all products","operationId": "listAll","responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } } } } },"/api/v1/fresh-products/orders/list": { "get": { "tags": [ "Purchase orders" ],"summary": "Get purchase order products by category","operationId": "findById","parameters": [ { "name": "Authorization","in": "header","required": true,"schema": { "type": "string" } },{ "name": "id","in": "query","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ApiError" } } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } } } } },"/api/v1/fresh-products/list": { "get": { "tags": [ "Product catalog" ],"summary": "List products by category","operationId": "listByCategory","parameters": [ { "name": "category","in": "query","required": false,"schema": { "type": "string","enum": [ "FS","RF","FF" ],"default": "FS" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductResponseDTO" } } } } } } } },"/api/v1/fresh-products/list/stocks": { "get": { "tags": [ "Product catalog" ],"summary": "List stocks of a product by productCode","operationId": "listStockByProductCode","parameters": [ { "name": "Authorization","in": "header","required": true,"schema": { "type": "string" } },{ "name": "productCode","in": "query","required": true,"schema": { "type": "string" } },{ "name": "order","in": "query","required": false,"schema": { "type": "string","enum": [ "C","F","c","f" ],"default": "C" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductStockResponseDTO" } } } } } } } },"/api/v1/fresh-products/due-date": { "get": { "tags": [ "Product catalog" ],"summary": "List stocks of a product that their dueDate are contained in the date interval","operationId": "listStockByProductCode_1","parameters": [ { "name": "dayQuantity","in": "query","required": true,"schema": { "minimum": 0,"type": "integer","format": "int32" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductStockResponseDTO" } } } } } } } },"/api/v1/fresh-products/due-date/list": { "get": { "tags": [ "Product catalog" ],"summary": "List stocks of a product that their dueDate are contained in the date interval and filtered by product category","operationId": "listStockByProductCode_2","parameters": [ { "name": "dayQuantity","in": "query","required": true,"schema": { "minimum": 0,"type": "integer","format": "int32" } },{ "name": "productCategory","in": "query","required": true,"schema": { "type": "string","enum": [ "FS","RF","FF" ] } },{ "name": "order","in": "query","required": false,"schema": { "type": "string","enum": [ "asc","desc" ],"default": "asc" } } ],"responses": { "400": { "description": "Bad Request","content": { "*/*": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ApiError" },{ "type": "array","items": { "$ref": "#/components/schemas/ValidationError" } } ] } } } },"500": { "description": "Internal Server Error","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"404": { "description": "Not Found","content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"401": { "description": "Unauthorized","content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },"200": { "description": "OK","content": { "application/json": { "schema": { "type": "array","items": { "$ref": "#/components/schemas/ProductStockResponseDTO" } } } } } } } } },"components": { "schemas": { "ValidationError": { "type": "object","properties": { "field": { "type": "string" },"message": { "type": "string" } } },"ApiError": { "type": "object","properties": { "error": { "type": "string" },"message": { "type": "string" },"status": { "type": "integer","format": "int32" } } },"ProductsDTO": { "required": [ "productId","quantity" ],"type": "object","properties": { "productId": { "type": "string" },"quantity": { "minimum": 0,"type": "integer","format": "int32" } } },"PurchaseOrderDTO": { "required": [ "buyerId","date","id","products","statusCode" ],"type": "object","properties": { "id": { "type": "integer","format": "int64" },"date": { "type": "string","format": "date" },"buyerId": { "type": "string" },"statusCode": { "type": "string" },"products": { "type": "array","items": { "$ref": "#/components/schemas/ProductsDTO" } } } },"InboundOrderResponseDTO": { "type": "object","properties": { "batchStock": { "type": "array","items": { "$ref": "#/components/schemas/StockDTO" } } } },"LocalTime": { "type": "object","properties": { "hour": { "type": "integer","format": "int32" },"minute": { "type": "integer","format": "int32" },"second": { "type": "integer","format": "int32" },"nano": { "type": "integer","format": "int32" } } },"StockDTO": { "required": [ "batchNumber","currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","price","productCode" ],"type": "object","properties": { "batchNumber": { "type": "integer","format": "int32" },"productCode": { "type": "string" },"currentTemperature": { "type": "number","format": "double" },"minimumTemperature": { "type": "number","format": "double" },"initialQuantity": { "minimum": 0,"type": "integer","format": "int32" },"currentQuantity": { "minimum": 0,"type": "integer","format": "int32" },"price": { "type": "number","format": "double" },"manufacturingDate": { "type": "string","format": "date" },"manufacturingTime": { "$ref": "#/components/schemas/LocalTime" },"dueDate": { "type": "string","format": "date" } } },"InboundOrderDTO": { "required": [ "batchStock","orderDate","orderNumber","section" ],"type": "object","properties": { "orderNumber": { "type": "integer","format": "int64" },"orderDate": { "type": "string","format": "date" },"section": { "$ref": "#/components/schemas/SectionDTO" },"batchStock": { "type": "array","items": { "$ref": "#/components/schemas/StockDTO" } } } },"SectionDTO": { "required": [ "sectionCode","warehouseCode" ],"type": "object","properties": { "sectionCode": { "type": "string" },"warehouseCode": { "type": "string" } } },"AccountResponseDTO": { "type": "object","properties": { "username": { "type": "string" },"password": { "type": "string" },"token": { "type": "string" } } },"CountryHouseDTO": { "required": [ "country","name" ],"type": "object","properties": { "name": { "maxLength": 2147483647,"minLength": 3,"type": "string" },"country": { "maxLength": 2147483647,"minLength": 3,"type": "string" } } },"ProductCategoryResponseDTO": { "type": "object","properties": { "categoryCode": { "type": "string" },"id": { "type": "integer","format": "int64" } } },"ProductResponseDTO": { "type": "object","properties": { "id": { "type": "integer","format": "int64" },"productCode": { "type": "string" },"minimumTemperature": { "type": "number","format": "double" },"manufacturingDate": { "type": "string","format": "date" },"manufacturingTime": { "$ref": "#/components/schemas/LocalTime" },"dueDate": { "type": "string","format": "date" },"productCategory": { "$ref": "#/components/schemas/ProductCategoryResponseDTO" },"price": { "type": "number","format": "double" } } },"IInfoStockDTO": { "type": "object","properties": { "batchNumber": { "type": "integer","format": "int32" },"currentQuantity": { "type": "integer","format": "int32" },"sectionCode": { "type": "string" },"warehouseCode": { "type": "string" },"dueDate": { "type": "string","format": "date" } } },"ProductStockResponseDTO": { "type": "object","properties": { "productId": { "type": "string" },"batchStock": { "type": "array","items": { "$ref": "#/components/schemas/IInfoStockDTO" } } } } },"securitySchemes": { "bearerAuth": { "type": "http","name": "bearerAuth","scheme": "bearer","bearerFormat": "JWT" } } } }